Implemented azScan

+++ src/edu/ucsb/deepspace/TelescopeInterface.java
@@ -23,7 +23,8 @@ public interface TelescopeInterface {
 	public double getOffset(Axis axis);
 	
 	public void calibrate(Coordinate c);
-	public void rasterScan(double minAz, double maxAz, double minEl, double maxEl, double reps);
+	public void rasterScan(double minAz, double maxAz, double minEl, double maxEl, int reps);
+
